feat: remove TurboDataItemVerifier
We actually do not need to worry about validating the signatures of signed data items sent to the SDK. They will just be forwarded to the Upload Service and it will be responsible for validating the signatures of those data items. Makes it much easier on the SDK side and avoids having to worry about verifying sigs in web vs. node environments.