fix(axios): set `maxRedirects` to `0` on our axios instances
We found a bug that was causing large amount of backpressure when uploading larger files (3GiB+) via the SDK. Axios loads the stream into memory, and backpressure is created while posting to `https`. We need to look further into any unintended consequeneces of this change, but it does allow larger files to be uploaded.