fix(deps): include bitcoinjs-lib as a dependency
Ideally, this is just a peer dependency and does not impact downstream clients if they do not need it, but nextjs struggles to tree-shake this dep out as the `@kepler-wallet` deps require it as a peer depedency. In the future we should look how to extract the kyve implementation so it can be optionally added as a plugin/separte dep and only used when necessary.