fix(uploader): do not resume stream internally, leave that to caller
This is a situation in which Readables start flowing as soon as .resume() is called if there is either a data handler or pipe setup for the stream. Because we create the data internally, when we call .resume() we immediately let bytes flow. This means the post request for the data item actually mat not get all the bytes, causing it to upload an unexpected data. This changes add returning a `resume()` function which resumes the original stream when called, and allows you to set up consumer promises before bytes start flowing.